+++ lib/src/service/tweets/exclude_tweet_type.dart
@@ -2,7 +2,7 @@
 // Redistribution and use in source and binary forms, with or without
 // modification, are permitted provided the conditions.
 
-enum TweetExcludeType {
+enum ExcludeTweetType {
   retweets,
   replies,
 }