+++ CHANGELOG.md
@@ -50,6 +50,7 @@
 - Added `sinceTweetId` and `untilTweetId` parameters to `TweetsService.lookupTweets`. ([#309](https://github.com/twitter-dart/twitter-api-v2/issues/309))
 - Added `excludes` parameter to `TweetsService.lookupTweets`. ([#308](https://github.com/twitter-dart/twitter-api-v2/issues/308))
 - Added `sinceTweetId` and `untilTweetId` parameters to `lookupHomeTimeline`. ([#307](https://github.com/twitter-dart/twitter-api-v2/issues/307))
+- Added `excludes` parameter to `TweetsService.lookupHomeTimeline`. ([#306](https://github.com/twitter-dart/twitter-api-v2/issues/306))
 
 ## v2.7.0